Return correct light count from GenerateSourceLightBuffers
The only functional change is fixing the light count calculation in GenerateSourceLightBuffers.  It actually returns the end index for the probe portion of the light list, not the count of probes.  This doesn't lead to any real bugs because the entries copied into the light data buffers are zero'd out because C# is nice to you.  This would likely manifest in bugs in C/C++.

There are also a lot of readability changes made to GenerateSourceLightBuffers.